+++ src/Makefile
@@ -48,7 +48,7 @@ ifndef BUILD_STATIC_ONLY
 	VERSION_SCRIPT := libbpf.map
 endif
 
-HEADERS := bpf.h libbpf.h btf.h xsk.h libbpf_util.h \
+HEADERS := bpf.h libbpf.h btf.h xsk.h \
 	   bpf_helpers.h bpf_helper_defs.h bpf_tracing.h \
 	   bpf_endian.h bpf_core_read.h libbpf_common.h
 UAPI_HEADERS := $(addprefix $(TOPDIR)/include/uapi/linux/,\
